+++ packages/dian/src/Soap/SoapMessageIdGenerator.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tribux\Dian\Soap;
+
+interface SoapMessageIdGenerator
+{
+    public function generate(): string;
+}